Revert "FROMLIST: vb2: Move cache synchronisation from buffer done to dqbuf handler"
This reverts commit 13ef5fcc0273f6b231d15584f747fdce61fb4056.

The original reason for landing this work in progress patch was to speed
up JPEG decoding on Elm, where we needed a software format conversion of
the decoder output to make it usable with the rest of the stack and
which was slow if the memory mapping was not cacheable. In turn it made
the cache synchronization happen in interrupt context and affecting
performance badly.

However, there are 2 significant issues with that patch:
 - The work upstream did not continue, so it stayed FROMLIST forever and
   actually started conflicting and causing functional issues with further
   upstream cherry picks (e.g. Request API).
 - It breaks drivers which need to access the buffers from their vb2
   .buf_finish callbacks, since it moves the cache synchronization after
   the call to it.

Since there is no board using chromeos-4.19 and affected by the original
problem, let's revert and bring our downstream diffstat down.

BUG=b:126440061
BUG=chromium:938781
TEST=Run `cros_camera_test
 --camera_hal_path=/usr/lib/camera_hal/mtk_cam_hal.so --v=0
 --gtest_filter=*TakePictureTest*` on Kukui few times

Change-Id: Ibdedfa798c8c06d869726569c66098c11c0b8f22
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1505078
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>