Revert "FROMLIST: vb2: dma-sg: Let drivers decide DMA attrs of MMAP and USERPTR bufs"
This reverts commit f24b395e9e3ebeacd7f1627249413e5c4322505c.

The original reason for landing this work in progress patch was to speed
up JPEG decoding on Elm, where we needed a software format conversion of
the decoder output to make it usable with the rest of the stack and
which was slow if the memory mapping was not cacheable. In turn it made
the cache synchronization happen in interrupt context and affecting
performance badly.

However, there are 2 significant issues with that patch:
 - The work upstream did not continue, so it stayed FROMLIST forever and
   actually started conflicting and causing functional issues with further
   upstream cherry picks (e.g. Request API).
 - It breaks drivers which need to access the buffers from their vb2
   .buf_finish callbacks, since it moves the cache synchronization after
   the call to it.

Since there is no board using chromeos-4.19 and affected by the original
problem, let's revert and bring our downstream diffstat down.

BUG=chromium:938781
TEST=Pre-CQ passes

Change-Id: I26e2a0616e597a99b8d42c3440dfd6ba22956960
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1505079
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>