Revert "FROMLIST: vb2: Anticipate queue specific DMA attributes for USERPTR buffers"
This reverts commit b501ccaec3b0fa564615d6f68a02a3a83d0d96c7.

The original reason for landing this work in progress patch was to speed
up JPEG decoding on Elm, where we needed a software format conversion of
the decoder output to make it usable with the rest of the stack and
which was slow if the memory mapping was not cacheable. In turn it made
the cache synchronization happen in interrupt context and affecting
performance badly.

However, there are 2 significant issues with that patch:
 - The work upstream did not continue, so it stayed FROMLIST forever and
   actually started conflicting and causing functional issues with further
   upstream cherry picks (e.g. Request API).
 - It breaks drivers which need to access the buffers from their vb2
   .buf_finish callbacks, since it moves the cache synchronization after
   the call to it.

Since there is no board using chromeos-4.19 and affected by the original
problem, let's revert and bring our downstream diffstat down.

BUG=chromium:938781
TEST=Pre-CQ passes

Change-Id: I3ec551e661b324ded71238842cf63f88a5326da4
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1505080
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>