FROMGIT: dt-bindings: phy-qcom-qmp: Add UFS PHY reset
Add a required reset to the SDM845 UFS phy to express the PHY reset
bit inside the UFS controller register space. Before this change, this
reset was not expressed in the DT, and the driver utilized two different
callbacks (phy_init and phy_poweron) to implement a two-phase
initialization procedure that involved deasserting this reset between
init and poweron. This abused the two callbacks and diluted their
purpose.

That scheme does not work as regulators cannot be turned off in
phy_poweroff because they were turned on in init, rather than poweron.
The net result is that regulators are left on in suspend that shouldn't
be.

This new scheme gives the UFS reset to the PHY, so that it can fully
initialize itself in a single callback. We can then turn regulators on
during poweron and off during poweroff.

Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
(cherry picked from commit eb263a6e74176de180e182e8c8a7cd38fb78e3f4
 https://git.kernel.org/pub/scm/linux/kernel/git/kishon/linux-phy.git/ next)

BUG=b:112846263
TEST=Build kernel for Cheza and boot

Change-Id: Ifb31de28f3c691fe97e20404cc639166bfd329b2
Reviewed-on: https://chromium-review.googlesource.com/1541731
Commit-Ready: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>