FROMGIT: ARM: dts: rockchip: Add rk3288-veyron-jerry rev 10-15
As far as I can tell/remember rev10 was originally created to support
making a SKU of jerry that had a different LCD.  rev11-rev15 were
added to give some wiggle room for future builds.  Downstream has a
separate device tree for rev10-rev15 (compared to rev3-rev7) with the
expectation that differences relating to the LCD would be accounted
for there but nothing was ever added to the rev10-rev15 making it
identical to the rev3-rev7 one.

It's likely nothing actually shipped with rev10-rev15 but they are
listed in the downstream kernel's device tree and it seems like it
should add a little safety if we match them here just in case
something actually shipped with one of these revisions and that device
will break if we don't claim support.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
(cherry picked from commit 0c4cac5e8f0313a8777055c0c66a2216f78c6054
 git://git.kernel.org/pub/scm/linux/kernel/git/mmind/linux-rockchip.git v5.2-armsoc/dts32)

BUG=chromium:941638
TEST=Build and boot

Change-Id: If150f57560c65bc28970ff14e40e082da7d35519
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1536980
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Ryan Case <ryandcase@google.com>