CHROMIUM: iwl7000: parse command version TLV
In newer firmware images there's a command version TLV that
states, for each listed command or notification, which version
of the command and/or notification structure is used. Read and
keep this data to be able to use it in the future.

Change-Id: Id1e28113ff03a55c7979f25a651e8a2c26de12d4
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: cadb07fc20a2b11f4aec147160c88c397a27a0d7