CHROMIUM: wilco_ec: Add exclusive lock to event node
Add an exclusive lock to the wilco events node, so only
one user can open the file at a time. This is to guarantee
that the consumer receives every single event, since if there
were multiple consumers of the queue, then there would be a
race condition for which would receive each event.

Also fix in Documentation where the dev node appears.

BUG=b:129421695
TEST=only one client can open the node at a time.

Change-Id: Ic21399b486d29ec55315f4dc9e6fd256e2ee586f
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1541838