UPSTREAM: platform/chrome: cros_ec_debugfs: Remove dev_warn when console log is not supported
If the debugfs interface is enabled, every time a CrOS device is
instantiated a warning like this can appear for every probed device.

 "device does not support reading the console log"

The warning message adds nothing, rather it is source of confusion as
this is expected on some cases. For example, on Samus, that has a
cros-ec and a cros-pd instance the message appears twice, and I suspect
this will happen also on those devices that has a non-standard EC.

If the command is not supported just return silently and don't print the
warning, otherwise the code will already print an error.

BUG=None
TEST=Check on samus-kernelnext

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit e43c426a777eb2554d8a29d31311debe897badd9)
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>

Change-Id: I417db0a42060ecd0a6c727a9368f99ac91f72037
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1707103
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>