FROMLIST: Add Wilco EC keyboard backlight LEDs support
(am from https://lkml.org/lkml/2019/4/8/1062)

The EC is in charge of controlling the keyboard backlight on
the Wilco platform. We expose a standard LED class device at
/sys/class/leds/chromeos::kbd_backlight. This driver is modeled
after the standard Chrome OS keyboard backlight driver at
drivers/platform/chrome/cros_kbd_led_backlight.c

Some Wilco devices do not support a keyboard backlight. This
is checked via wilco_ec_keyboard_leds_exist() in the core driver,
and a platform_device will only be registered by the core if
a backlight is supported.

After an EC reset the backlight could be in a non-PWM mode.
Earlier in the boot sequence the BIOS should send a command to
the EC to set the brightness, so things **should** be set up,
but we double check in probe() as we query the initial brightness.
If not set up, then set the brightness to 0.

Since the EC will never change the backlight level of its own accord,
we don't need to implement a brightness_get() method.

There is still ongoing discussion upstream as to the desired name of
the LED. I went with what groeck@chromium.org and dtor@chromium.org
want. If we need to tweak this later to reflect what actually lands
upstream, it will be really easy and it won't affect how powerd is
finding the LED device.

BUG=b:123637710
TEST=on a device that supports the backlight:
localhost ~ # echo 60 >
/sys/class/leds/chromeos\:\:kbd_backlight/brightness
localhost ~ # cat /sys/class/leds/chromeos\:\:kbd_backlight/brightness
60
localhost ~ # cat
/sys/class/leds/chromeos\:\:kbd_backlight/max_brightness
100
On a device that doesn't support the backlight, the sysfs
entry does not exist.

Change-Id: I3b1498f2ee65d77f274f0dd58cf8c5ec6adf3ed5
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1512277