CHROMIUM: dt-bindings: soc: rockchip: Bindings for the rk3288 service bus
This device contains registers to control the memory scheduler on
rk3288 SoCs.  It is accessed during DDRFreq transitions and also at
suspend/resume.

NOTE that in the downstream Chrome OS 3.14 kernel this was called the
"noc" but I renamed it to "service bus" to match the TRM.  Downstream
Chrome OS didn't have any bindings for this.

BUG=chromium:948742
TEST=suspend_stress_test

Change-Id: Ica3eb2d9c0bcb036cda1e1c613ee7bac34f7b29e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1570550
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>