CHROMIUM: dt-bindings: sram: rockchip: Document ddr sram bindings
In order to change the frequency that SDRAM is running at we need a
bit of space (16K) in SRAM to contain the code for the frequency
change.

Therefore a reserved section sub-node has to be added to the mmio-sram
declaration.

Note that downstream Chrome OS 3.14 didn't have these bindings but
these bindings match what was actually used in Chrome OS.

BUG=chromium:948742
TEST=suspend_stress_test

Change-Id: Id309e7b041e8b054d015f9c6927cdff1b44f6b1b
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1570551
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>