CHROMIUM: config: Turn on CONFIG_CPU_BOOST for arm32
The input boost is super awesome and we definitely want it for arm32.

This is the equivalent of:

  echo "CONFIG_CPU_BOOST=y" >> chromeos/config/armel/common.config
  ./chromeos/scripts/kernelconfig olddefconfig

This will allow boards to turn on input boosting by having an init
script that does something like:
  echo 75 > /sys/module/cpu_boost/parameters/input_boost_freq_percent

As an example:
  https://chromium.googlesource.com/chromiumos/overlays/board-overlays/+/
  refs/heads/master/
  baseboard-gru/chromeos-base/chromeos-bsp-baseboard-gru/files/platform-cpusets.conf

Note: this means we've got this on for arm32 and arm64 but not x86.  I
guess on x86 the BIOS magic handles the boost?

BUG=chromium:941638
TEST=Change dev_debug to dev_info and type; Set input_boost_freq_percent

Change-Id: Id65e635e6c543616e75cacc8c937828ce824e5c9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1559933
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>